fix(librofm): UserMetadata.{TrackIndex,TrackSeconds} are floats, not ints
Next field-shape surprise after the ISBN-as-number one: libro.fm
encodes user_metadata.track_seconds as a JSON float — even when the
value is zero (`0.0`), which Go's decoder refuses to fold into an
int field. Same applies to track_index in principle, even if we
haven't observed a fractional value there yet.

internal/librofm/types.go
- TrackIndex and TrackSeconds: int → float64.
- Comment notes the wire-format gotcha so the next reader doesn't
  re-tighten this on the basis of the field name alone.

internal/librofm/client_test.go
- TestLibrary_AcceptsFloatUserMetadataNumbers: mixed payload, one
  entry with `0.0` and one with `1234.567`, asserts both parse and
  the values round-trip.

UserMetadata isn't read by the reconciler in v1 — it's preserved for
the deferred wishlist/progress sync. Could've been replaced with
json.RawMessage to side-step shape questions entirely, but the typed
fields are easier to discover when that work eventually lands. Trade
documented-shape value for slightly-fragile decoding; relax fields
as they bite.

+++ internal/librofm/types.go
@@ -80,12 +80,16 @@ type PdfExtra struct {
 
 // UserMetadata is the per-user info present in /api/v10/library entries.
 // We don't act on most fields in v1 but keep them for future progress sync.
+//
+// Note: libro.fm encodes the numeric fields here as JSON floats (e.g.
+// `"track_seconds": 0.0`) even when the value has no fractional component.
+// Use float64 — Go's decoder rejects a fractional value into an int.
 type UserMetadata struct {
-	Finished     bool   `json:"finished,omitempty"`
-	TrackIndex   int    `json:"track_index,omitempty"`
-	TrackSeconds int    `json:"track_seconds,omitempty"`
-	AddedAt      string `json:"added_at,omitempty"`
-	Hidden       bool   `json:"hidden,omitempty"`
+	Finished     bool    `json:"finished,omitempty"`
+	TrackIndex   float64 `json:"track_index,omitempty"`
+	TrackSeconds float64 `json:"track_seconds,omitempty"`
+	AddedAt      string  `json:"added_at,omitempty"`
+	Hidden       bool    `json:"hidden,omitempty"`
 }
 
 // LibraryPage is one page of /api/v10/library.
